# openapi.yaml
openapi: 3.0.1
info:
  title: Mentor-Mentee Matching API
  description: API for matching mentors and mentees in a mentoring platform
  version: 1.0.0
  contact:
    name: Mentor-Mentee Matching App
  license:
    name: MIT
servers:
  - url: http://localhost:8088/api
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /me:
    get:
      operationId: getCurrentUser
      tags:
        - User Profile
      summary: Get current user information
      description: Retrieve the profile information of the currently authenticated user
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/MentorProfile'
                  - $ref: '#/components/schemas/MenteeProfile'
        '401':
          description: Unauthorized - authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profile:
    put:
      operationId: updateProfile
      tags:
        - User Profile
      summary: Update user profile
      description: Update the profile information of the currently authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/UpdateMentorProfileRequest'
                - $ref: '#/components/schemas/UpdateMenteeProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/MentorProfile'
                  - $ref: '#/components/schemas/MenteeProfile'
        '400':
          description: Bad request - invalid payload format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /images/{role}/{id}:
    get:
      operationId: getProfileImage
      tags:
        - User Profile
      summary: Get profile image
      description: Retrieve the profile image for a specific user
      parameters:
        - name: role
          in: path
          required: true
          schema:
            type: string
            enum: [mentor, mentee]
          description: User role (mentor or mentee)
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        '200':
          description: Profile image retrieved successfully
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized - authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    MentorProfile:
      type: object
      required:
        - id
        - email
        - role
        - profile
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: "user@example.com"
        role:
          type: string
          enum: [mentor]
          example: "mentor"
        profile:
          $ref: '#/components/schemas/MentorProfileDetails'
    MenteeProfile:
      type: object
      required:
        - id
        - email
        - role
        - profile
      properties:
        id:
          type: integer
          example: 10
        email:
          type: string
          format: email
          example: "user@example.com"
        role:
          type: string
          enum: [mentee]
          example: "mentee"
        profile:
          $ref: '#/components/schemas/MenteeProfileDetails'
    MentorProfileDetails:
      type: object
      required:
        - name
        - bio
        - imageUrl
        - skills
      properties:
        name:
          type: string
          example: "Alice"
        bio:
          type: string
          example: "Frontend mentor"
        imageUrl:
          type: string
          example: "/images/mentor/1"
        skills:
          type: array
          items:
            type: string
          example: ["React", "Vue"]
    MenteeProfileDetails:
      type: object
      required:
        - name
        - bio
        - imageUrl
      properties:
        name:
          type: string
          example: "Alice"
        bio:
          type: string
          example: "Frontend mentee"
        imageUrl:
          type: string
          example: "/images/mentee/10"
    UpdateMentorProfileRequest:
      type: object
      required:
        - id
        - name
        - role
        - bio
        - image
        - skills
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Alice"
        role:
          type: string
          enum: [mentor]
          example: "mentor"
        bio:
          type: string
          example: "Frontend mentor"
        image:
          type: string
          format: byte
          description: "Base64 encoded image string"
        skills:
          type: array
          items:
            type: string
          example: ["React", "Vue"]
    UpdateMenteeProfileRequest:
      type: object
      required:
        - id
        - name
        - role
        - bio
        - image
      properties:
        id:
          type: integer
          example: 21
        name:
          type: string
          example: "Alice"
        role:
          type: string
          enum: [mentee]
          example: "mentee"
        bio:
          type: string
          example: "Frontend mentee"
        image:
          type: string
          format: byte
          description: "Base64 encoded image string"
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Error message"
        details:
          type: string
          example: "Detailed error information"
...